First Commit A00001.1
A minimal functioning version of the games functionality so far. Only
graphic place holders will be used until all functionality is complete.
Visual Studio 2013 is required for easy compiling.
This game requires the latest version of the SFML library.
The Entity class controls and makes the sprite setup easy and is made to
be used with any game.